requirement.txt
azure-cognitiveservices-speech
pytube
fpdf
streamlit